clean up metrics
