+++ protocols/substrate/components/metrics/vars.go
@@ -1,0 +1,3 @@
+package metrics
+
+var EncodingVersion uint8 = 1
